Restructure directory layout to support UI integration
