+++ KeyRemapper/Installers/GameplayInstaller.cs
@@ -1,0 +1,12 @@
+﻿using Zenject;
+using KeyRemapper.Logic;
+
+namespace KeyRemapper.Installers;
+
+public class GameplayInstaller : Installer
+{
+    public override void InstallBindings()
+    {
+        Container.BindInterfacesAndSelfTo<OpenXRPausePatch>().AsSingle();
+    }
+}
